+++ src/db/dao/CardDAO.java
@@ -4,7 +4,9 @@ import db.entities.Card;
 
 import java.util.ArrayList;
 import java.util.List;
+
 import db.jpa.CardJPA;
+
 import javax.persistence.EntityManagerFactory;
 import javax.persistence.Persistence;
 
@@ -13,13 +15,14 @@ public class CardDAO implements DAO<Card> {
     private final List<Card> cards = new ArrayList<>();
     public EntityManagerFactory emf;
     private final CardJPA cardJPA;
+
     public CardDAO() {
-        emf= Persistence.createEntityManagerFactory("AL2000");
+        emf = Persistence.createEntityManagerFactory("AL2000");
         cardJPA = new CardJPA(emf.createEntityManager());
     }
 
     @Override
-    public Card get(long id) {
+    public Card get(int id) {
         return cardJPA.get(id);
     }
